backend/requirements.txt
asgiref==3.8.1
dj-database-url==2.3.0
Django==5.2.1
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
filelock==3.18.0
fsspec==2025.5.0
gunicorn==23.0.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
packaging==25.0
pillow==11.2.1
psycopg2==2.9.10
psycopg2-binary==2.9.10
PyJWT==2.9.0
setuptools==80.8.0
sqlparse==0.5.3
sympy==1.14.0
torch==2.7.0
torchvision==0.22.0
typing_extensions==4.13.2
tzdata==2025.2
whitenoise==6.9.0
